requirements.txt
crewai==0.51.1
openai>=1.13.3
fastapi 
uvicorn[standard] 
pydantic
crewai-tools
langchain 
chromadb 
sentence-transformers
pdfplumber 
pymupdf 
python-docx 
pypdf
beautifulsoup4 
httpx 
lxml
scikit-learn 
numpy 
pandas
apscheduler
python-dotenv